@@ -1,56 +1,64 @@
 {
   "name": "candifood-ui",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
+    "build-prod": "ng build --prod --aot",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "candifood-core": "ng build @candifood/core --prod"
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "1.0.0-beta.2",
-    "@angular/animations": "5.1.x",
-    "@angular/common": "5.1.x",
-    "@angular/compiler": "5.1.x",
-    "@angular/core": "5.1.x",
-    "@angular/forms": "5.1.x",
-    "@angular/http": "5.1.x",
-    "@angular/platform-browser": "5.1.x",
-    "@angular/platform-browser-dynamic": "5.1.x",
-    "@angular/router": "5.1.x",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-beta.8",
-    "angularfire2": "5.0.0-rc.5-next",
-    "bootstrap": "4.0.0-beta.2",
-    "core-js": "2.5.3",
-    "crypto-js": "3.1.9-1",
-    "firebase": "4.8.1",
-    "font-awesome": "4.7.0",
-    "rxjs": "5.5.5",
-    "zone.js": "0.8.18"
+    "@agm/core": "1.0.0-beta.3",
+    "@angular/animations": "6.0.2",
+    "@angular/common": "6.0.2",
+    "@angular/compiler": "6.0.2",
+    "@angular/core": "6.0.2",
+    "@angular/forms": "6.0.2",
+    "@angular/http": "6.0.2",
+    "@angular/platform-browser": "6.0.2",
+    "@angular/platform-browser-dynamic": "6.0.2",
+    "@angular/router": "6.0.2",
+    "@candifood/core": "0.0.9",
+    "@fortawesome/angular-fontawesome": "0.1.0-9",
+    "@fortawesome/fontawesome-svg-core": "^1.2.0-11",
+    "@fortawesome/free-regular-svg-icons": "^5.1.0-8",
+    "@fortawesome/free-solid-svg-icons": "^5.1.0-8",
+    "@fortawesome/free-brands-svg-icons": "5.1.0-8",
+    "@ng-bootstrap/ng-bootstrap": "2.0.0",
+    "@ng-bootstrap/schematics": "2.0.0-alpha.1",
+    "bootstrap": "4.1.1",
+    "core-js": "2.5.4",
+    "rxjs": "6.1.0",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.3",
-    "@angular/compiler-cli": "5.1.x",
-    "@angular/language-service": "5.1.x",
-    "@types/jasmine": "2.8.2",
+    "@angular/compiler-cli": "6.0.2",
+    "@angular-devkit/build-ng-packagr": "0.6.1",
+    "@angular-devkit/build-angular": "0.6.1",
+    "ng-packagr": "3.0.0-rc.2",
+    "tsickle": ">=0.25.5",
+    "tslib": "1.7.1",
+    "typescript": "2.7.2",
+    "@angular/cli": "6.0.1",
+    "@angular/language-service": "6.0.2",
+    "@types/jasmine": "2.8.6",
     "@types/jasminewd2": "2.0.3",
-    "@types/node": "6.0.60",
-    "codelyzer": "4.0.2",
-    "jasmine-core": "2.8.0",
+    "@types/node": "8.9.4",
+    "codelyzer": "4.2.1",
+    "jasmine-core": "2.99.1",
     "jasmine-spec-reporter": "4.2.1",
     "karma": "1.7.1",
     "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.3.0",
+    "karma-coverage-istanbul-reporter": "1.4.2",
     "karma-jasmine": "1.1.1",
     "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.2.2",
-    "ts-node": "3.2.0",
-    "tslint": "5.8.0",
-    "typescript": "2.6.2"
+    "protractor": "5.3.0",
+    "ts-node": "5.0.1",
+    "tslint": "5.9.1"
   }
 }
