@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {Routes, RouterModule} from '@angular/router';
+import {HomeComponent} from './home/components/home/home.component';
 import {DisplayRestaurantComponent} from './restaurant/components/display-restaurant/display-restaurant.component';
-import {LoginComponent} from './core/components/login/login.component';
 
 const routes: Routes = [
   {
@@ -9,13 +9,13 @@ const routes: Routes = [
     component: DisplayRestaurantComponent
   },
   {
-    path: 'login',
-    component: LoginComponent
+    path: 'display-restaurant',
+    component: DisplayRestaurantComponent
   }
-  ];
+];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes/*, {enableTracing: true}*/)],
   exports: [RouterModule]
 })
 export class AppRoutingModule {
