@@ -1,3 +1,12 @@
 export const environment = {
-  production: true
+  production: true,
+  restUrl: 'https://andycandifood.appspot.com',
+  trackAnalytics: true,
+  useBasicAuth: false,
+  alertDelayInSeconds: 7,
+  'ga': {
+    'key': 'tempPwd',
+    'iv': 'XXXX',
+    'trackingId': 'UA-115890187-1'
+  }
 };
