@@ -1,8 +1,26 @@
-// The file contents for the current environment will overwrite these during build.
-// The build system defaults to the dev environment which uses `environment.ts`, but if you do
-// `ng build --env=prod` then `environment.prod.ts` will be used instead.
-// The list of which env maps to which file can be found in `.angular-cli.json`.
+// This file can be replaced during build by using the `fileReplacements` array.
+// `ng build ---prod` replaces `environment.ts` with `environment.prod.ts`.
+// The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  // restUrl: 'http://localhost:8080',
+  restUrl: 'https://andycandifood.appspot.com',
+  trackAnalytics: true,
+  useBasicAuth: false,
+  alertDelayInSeconds: 7,
+  'ga': {
+    'key': 'tempPwd',
+    'iv': 'XXXX',
+    'trackingId': 'UA-115890187-1'
+  }
 };
+
+
+/*
+ * In development mode, to ignore zone related error stack frames such as
+ * `zone.run`, `zoneDelegate.invokeTask` for easier debugging, you can
+ * import the following file, but please comment it out in production mode
+ * because it will have performance impact when throw error
+ */
+// import 'zone.js/dist/zone-error';  // Included with Angular CLI.
